+++ spring/src/main/java/io/cloudevents/spring/http/CloudEventHttpUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.cloudevents.spring.webmvc;
+package io.cloudevents.spring.http;
 
 import java.util.HashMap;
 import java.util.Map;
