+++ spring/src/main/java/io/cloudevents/spring/http/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Provides classes related to working with Cloud Events within the context of Spring MVC.
  */
-package io.cloudevents.spring.webmvc;
+package io.cloudevents.spring.http;
